Renamed namespace and organized classes
