+++ src/Outkeep.Application.Azure/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Outkeep.Caching;
 using Outkeep.Core;
-using Outkeep.Grains;
 using Serilog;
 using System;
 using System.Diagnostics.CodeAnalysis;
