+++ src/Outkeep.Application.Standalone/Program.cs
@@ -2,9 +2,9 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Orleans.Statistics;
+using Outkeep.Caching;
 using Outkeep.Core;
 using Outkeep.Core.Caching;
-using Outkeep.Grains;
 using Serilog;
 using System;
 using System.Diagnostics.CodeAnalysis;
