--- src/Outkeep.Grains/CacheGrainContextServiceCollectionExtensions.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace Outkeep.Grains
-{
-    public static class CacheGrainContextServiceCollectionExtensions
-    {
-        public static IServiceCollection AddCacheGrainContext(this IServiceCollection services)
-        {
-            return services.AddSingleton<ICacheGrainContext, CacheGrainContext>();
-        }
-    }
-}
