+++ src/Outkeep.Grains/Caching/CacheGrainOptions.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace Outkeep.Grains
+namespace Outkeep.Caching
 {
     public class CacheGrainOptions
     {
-        [Range(typeof(TimeSpan), "00:00:01", "00:00:29")]
+        [Range(typeof(TimeSpan), "00:00:01", "00:00:25")]
         public TimeSpan ReactivePollingTimeout { get; set; } = TimeSpan.FromSeconds(20);
 
         [Range(typeof(TimeSpan), "00:00:01", "24:00:00")]
