+++ src/Outkeep.Grains/Outkeep.Grains.csproj
@@ -11,6 +11,7 @@
     <PackageProjectUrl>https://github.com/outcomputelabs/outkeep</PackageProjectUrl>
     <RepositoryUrl>https://github.com/outcomputelabs/outkeep</RepositoryUrl>
     <Nullable>enable</Nullable>
+    <RootNamespace>Outkeep</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
